CI: there's no point of testing just the "db/session.py" file with the rest changes, so we just test the (new) "services/events.py" file.
